added postmsg page, verify page has broken js-ipfs
