+++ lib/backend/twitter.ts
@@ -58,7 +58,7 @@ async function postTweet(message: string) {
   if (resp.statusCode !== 201) {
     throw new Error("error posting tweet");
   } else {
-    const tweetID = resp["body"]["data"]["id"];
+    const tweetID = resp.body.data.id;
     const tweetURL = `https://twitter.com/${twitterAccount}/status/${tweetID}`;
     return tweetURL;
   }
