+++ package.json
@@ -17,8 +17,9 @@
     "ethers": "5.6.8",
     "got": "12.1.0",
     "heroicons": "^1.0.6",
-    "ipfs-core": "0.15.1",
+    "ipfs-core": "^0.15.2",
     "ipfs-http-client": "57.0.1",
+    "ipfs-repo": "^14.0.1",
     "localforage": "^1.10.0",
     "needle": "3.1.0",
     "next": "12.1.6",
