+++ pages/group/[groupId].tsx
@@ -1,0 +1,11 @@
+import {useRouter} from 'next/router';
+
+const GroupPage = () => {
+    const router = useRouter();
+    const { groupId } = router.query;
+
+
+    return;
+}
+
+export default GroupPage;
