+++ styles/globals.css
@@ -4,7 +4,7 @@
 
 body {
   background: black;
-  font-family: 'Space Mono';
+  font-family: "Space Mono";
   /* color: white; */
   font-size: 14px;
 }
